Update go.mod
pkg.go.dev read this when there was no LICENSE.txt in the repo.
Because the license was missing, pkg.go.dev refuses to display the
documentation. I added LICENSE.txt later but it won't update the
published package state until go.sum changes (I think). This is a test
to see if it triggers an update of the pkg.go.dev/zrhmn/argv page.